+++ src/modules/reservations/service/intention/info-reservation.strategy.ts
@@ -1,10 +1,9 @@
 import { Injectable, Logger } from "@nestjs/common";
-import { DeleteReservation, Intention, MultipleMessagesResponse, RoleEnum } from "src/lib";
+import { AddMissingFieldInput, DeleteReservation, Intention, MultipleMessagesResponse, RoleEnum } from "src/lib";
 import { AiService } from "src/modules/ai/service/ai.service";
 import { DatesService } from "src/modules/dates/service/dates.service";
 import { CacheService } from "src/modules/cache-context/cache.service";
 import { IntentionStrategyInterface, StrategyResult } from "./intention-strategy.interface";
-import { getMissingFields } from "../helpers/get-missing-fields.helper";
 
 @Injectable()
 export class InfoReservationStrategy implements IntentionStrategyInterface {
@@ -18,12 +17,27 @@ export class InfoReservationStrategy implements IntentionStrategyInterface {
 
     async execute(aiResponse: MultipleMessagesResponse): Promise<StrategyResult> {
         const waId = '123456789'
+
+        const mockedData: AddMissingFieldInput = {
+            waId: '123456789',
+            values: { // TODO: remove this mock once we receive the phone number from the user
+                phone: '1122334455',
+                date: aiResponse.date,
+                time: aiResponse.time,
+            },
+            messageSid: '123',
+        }
+
+        const history = await this.cacheService.getHistory(mockedData.waId);
+
+
         if (aiResponse.date && aiResponse.time) {
             // check exact datetime
+            // return if datetime is available or near datetime
         } else if (!aiResponse.date && !aiResponse.time) {
             // ask for date and time
         }
-        // check date, return a range of times
+        // datetime not available
 
         return { reply: 'response' };
     }